Reorganize PagedNode
