+++ src/osgEarth/PagedNode
@@ -1,0 +1,50 @@
+/* -*-c++-*- */
+/* osgEarth - Dynamic map generation toolkit for OpenSceneGraph
+* Copyright 2016 Pelican Mapping
+* http://osgearth.org
+*
+* osgEarth is free software; you can redistribute it and/or modify
+* it under the terms of the GNU Lesser General Public License as published by
+* the Free Software Foundation; either version 2 of the License, or
+* (at your option) any later version.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+* FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+* IN THE SOFTWARE.
+*
+* You should have received a copy of the GNU Lesser General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>
+*/
+#ifndef OSGEARTHUTIL_PAGEDNODE_H
+#define OSGEARTHUTIL_PAGEDNODE_H
+
+#include <osgEarth/Common>
+#include <osg/PagedLOD>
+
+namespace osgEarth
+{
+    class OSGEARTH_EXPORT PagedNode : public osg::Group
+    {
+    public:
+        PagedNode();
+
+        virtual osg::Node* loadChildren();
+
+        virtual void build();
+
+        virtual osg::BoundingSphere getKeyBound() const;
+
+        virtual bool hasChildren() const;
+
+        void setupPaging();
+
+        osg::Group* _attachPoint;
+        osg::PagedLOD* _plod;
+    };    
+}
+
+#endif // OSGEARTHUTIL_PAGEDNODE_H
